requirements.txt
accelerate==1.10.1
aiofiles==25.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
backoff==2.2.1
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
effdet==0.4.1
emoji==2.15.0
exa-py==1.16.1
fastapi==0.119.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
kiwisolver==1.4.9
langdetect==1.0.9
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.1
openai==2.6.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pgvector==0.4.1
pi_heif==1.1.1
pikepdf==9.11.0
pillow==12.0.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.1
pypdfium2==4.30.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
webencodings==0.5.1
wrapt==1.17.3
